labs/iot/gps-lab/code/pi/requirements.txt
azure-iot-device
python-dotenv
gps